Create DescribeImage.m
This is what we were working up to; it extracts the topological info using all the other functions and converts it to lines of whatever. It's unfinished right now. Use the image name as input, the output will be the lines of 'code' (just numbers and symbols right now though).